working solution
used variables set through nb_setval and b_setval to keep track of state, passed them through function arguments, performed appropriate state checks such as if current path cost was lesser than shortestPathCost so far. If yes, shortestPathCost was updated. shortest path like this was updated multiple times, and due to a lack of global variables, I was not able to find a way to print just the last occurrence --- i.e. just the final shortest path. Temporarily, I used asserta/1 to add successive occurrences as predicates one above the other, and simply called that predicate in the terminal to get the answer.